+++ src/Machine.cpp
@@ -7,6 +7,6 @@ Machine::Machine() : Product() {
     product_type = Product::ProductType::TYPE_MACHINE;
     static std::uniform_int_distribution<>
         machine_lifetime_dist(MACHINE_LIFETIME_MIN, MACHINE_LIFETIME_MAX);
-    lifetime = machine_lifetime_dist(Sim::get_random_generator());
+    lifetime = static_cast<double>(machine_lifetime_dist(Sim::get_random_generator()));
 }
